Navigation almost works
I managed to get navigation to work from the home fragment to the box fragment using a listener in the fragments that performs the action from the nav graph inside the fragment when the button is pushed.

Known bug: Curently crashes when going into box from home, pressing back (to home), and then pressing box again.