Relocated the database call to viewmodel
Contacts fragment sends the list from viewmodel to the adapter which fills the recyclerview. When an item is clicked a index is collected and sent to user to user fragment which uses the same viewmodel and collects a single contact which is then displayed. Next is to implement the remove function.